# PERF: pandas.core.sorting.compress_group_index for already sorted values