@mroeschke 2023-06-23
Nice find thanks @lukemanley 

Is there a further fastpath if `group_index` is completely unique (like a `np.arange` array)?